+++ apiapp/go.mod
@@ -1,0 +1,3 @@
+module apiapp
+
+go 1.21.2
