+++ apiapp/product/product.go
@@ -1,0 +1,6 @@
+package product
+
+type Product struct {
+	ID    int    `json:"id"`
+	Title string `json:"title"`
+}
