+++ apiapp/repository/interface.go
@@ -1,0 +1,5 @@
+package repository
+
+type ProductRepository interface {
+	Execute(productId int)
+}
